% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2_function.R
\name{countData}
\alias{countData}
\title{Assignment 2 Statistical Analysis Function}
\usage{
countData(countTable)
}
\arguments{
\item{countTable, sampleData}{The name/path of the file in quotes marks}
}
\value{
Count data table
}
\description{
The following package will input an RNA-seq count data and sample data into R.
}
\examples{
count_data <- countData("E-MTAB-2523.counts.txt");

}
