% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2_function.R
\name{dgeObj}
\alias{dgeObj}
\title{DGE function}
\usage{
dgeObj(countTable)
}
\arguments{
\item{countTable}{countTable after filtering}
}
\value{
dge table
}
\description{
Create DGE table
}
\examples{
dge <- dge(countTable);
}
