% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2_function.R
\name{filtering}
\alias{filtering}
\title{Performing filtering}
\usage{
filtering(countTable)
}
\arguments{
\item{countTable}{Filter count data}
}
\value{
Count table with low filtered values removed
}
\description{
Filters by rowmeans(log2(cpm())+1)
}
\examples{
filterdata <- filtering(countTable);
}
