% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2_function.R
\name{linmodel}
\alias{linmodel}
\title{Linear Models}
\usage{
linmodel(dge, desMatrix)
}
\arguments{
\item{dge}{dge variable and variable for the design matrix}
}
\value{
Data files and png files for enrichment analysis visual representation
}
\description{
Makes design matrix and Fits linear models and performs stat testing
}
\examples{
linearmodel <- linmodel(dge);
}
