% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2_function.R
\name{sampleData}
\alias{sampleData}
\title{Sample data reading}
\usage{
sampleData(sampleData)
}
\arguments{
\item{sampleData}{The name/path to the sample data file}
}
\value{
The sample data and the factors
}
\description{
Reads in the data of the sample table and creates factors for the information
}
\examples{
sample_data <- sampleData("E-MTAB-2523_sample table.txt");
}
